% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGAMax.R
\name{IGAMax}
\alias{IGAMax}
\title{Find test statistic value from a given data and corresponding critical value}
\usage{
IGAMax(data, k, alpha)
}
\arguments{
\item{data}{a real matrix}

\item{k}{a positive integer}

\item{alpha}{real number between 0 and 1 called significance level}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=4;N=c(20,10,10,30);mu=c(1,1,1,1);l=c(4,5,5,6)
g=NULL
for(i in 1:k)
{
 g[[i]]=mgcv::rig(N[i],mu[i],1/l[i])
}
data=g
IGAMax(data,k,0.05)
}
