% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGEst.R
\name{IGEst}
\alias{IGEst}
\title{Find estimates of means and scale-like parameters when there is a prior information that means are under order restrictions}
\usage{
IGEst(data, k)
}
\arguments{
\item{data}{a real data frame}

\item{k}{a positive integer}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
k=4;N=c(20,10,10,30);mu=c(1,1.2,1.3,1.5);l=c(4,5,5,6)
g=NULL
for(i in 1:k)
{
 g[[i]]=mgcv::rig(N[i],mu[i],1/l[i])
}
data=g
IGEst(data,k)
}
